Create README.md
Detailed to solve the following questions;
QUESTIONS

FizzBuzz
Write a program that prints the numbers from 1 to 100. For multiples of 3, print "Fizz"; for
multiples of 5, print "Buzz"; and for numbers that are multiples of both 3 and 5, print
"FizzBuzz". Fibonacci Sequence
Write a program to generate the Fibonacci sequence up to 100.


Power of Two
Write a program that takes an integer as input and returns true if the input is a power of two.

Examples:
8=> returns true
6=> returns false

Capitalize Words
Write a program that accepts a string as input, capitalizes the first letter of each word in the
string, and then returns the result string.
Examples:

"hi"=> returns "Hi"
"i love programming"=> returns "I Love Programming"

Reverse Integer

Write a program that takes an integer as input and returns an integer with reversed digit
ordering.
Examples:

For input 500, the program should return 5.
For input -56, the program should return -65.
For input -90, the program should return -9.
For input 91, the program should return 19.

 Count Vowels
Write a program that counts the number of vowels in a sentence.